Consolidate Pokemon types
